# api/docs/swagger.yaml
definitions:
...
  models.CreateStaff:
    properties:
      balance:
        type: number
      branch_id:
        type: string
      name:
        type: string
      staff_type:
        type: string
      tariff_id:
        type: string
    type: object
...
  models.CreateTransaction:
    properties:
      amount:
        type: integer
      sale_id:
        type: string
      source_type:
        type: string
      staff_id:
        type: string
      text:
        type: string
      type:
        type: string
    type: object
...
  models.GetAllStaff:
    properties:
      count:
        type: integer
      staffs:
        items:
          $ref: '#/definitions/models.Staff'
        type: array
    type: object
...
  models.GetAllTransactionResponse:
    properties:
      count:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/models.Transaction'
        type: array
    type: object
...
  models.Staff:
    properties:
      balance:
        type: number
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      staff_type:
        type: string
      tariff_id:
        type: string
      updated_at:
        type: string
    type: object
...
  models.Transaction:
    properties:
      amount:
        type: integer
      created_at:
        type: string
      id:
        type: string
      sale_id:
        type: string
      source_type:
        type: string
      staff_id:
        type: string
      text:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  response.ErrorResp:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /staff:
    get:
      consumes:
      - application/json
      description: gets all staffs based on limit, page and search by name
      parameters:
      - default: 10
        description: limit
        in: query
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: page
        in: query
        minimum: 1
        name: page
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllStaff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET  ALL STAFFS
      tags:
      - STAFF
    post:
      consumes:
      - application/json
      description: CREATES STAFF BASED ON GIVEN DATA
      parameters:
      - description: staff data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: CREATES STAFF
      tags:
      - STAFF
  /staff/{id}:
    delete:
      consumes:
      - application/json
      description: DELETES STAFF BASED ON ID
      parameters:
      - description: id of staff
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: DELETE STAFF BY ID
      tags:
      - STAFF
    get:
      consumes:
      - application/json
      description: get staff by ID
      parameters:
      - description: Staff ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Staff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET BY ID
      tags:
      - STAFF
    put:
      consumes:
      - application/json
      description: UPDATES STAFF BASED ON GIVEN DATA AND ID
      parameters:
      - description: id of staff
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: staff data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: UPDATE STAFF BY ID
      tags:
      - STAFF
...
  /transaction:
    get:
      consumes:
      - application/json
      description: gets all transaction based on limit, page and search by name
      parameters:
      - default: 10
        description: limit
        in: query
        minimum: 1
        name: limit
        type: integer
      - default: 1
        description: page
        in: query
        minimum: 1
        name: page
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAllTransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET  ALL TRANSACTION
      tags:
      - TRANSACTION
    post:
      consumes:
      - application/json
      description: CREATES TRANSACTION BASED ON GIVEN DATA
      parameters:
      - description: transaction data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: CREATES TRANSACTION
      tags:
      - TRANSACTION
  /transaction/{id}:
    delete:
      consumes:
      - application/json
      description: DELETES TRANSACTION BASED ON ID
      parameters:
      - description: id of transaction
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: DELETE TRANSACTION BY ID
      tags:
      - TRANSACTION
    get:
      consumes:
      - application/json
      description: get transaction by ID
      parameters:
      - description: Transaction ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: GET BY ID
      tags:
      - TRANSACTION
    put:
      consumes:
      - application/json
      description: UPDATES TRANSACTION BASED ON GIVEN DATA AND ID
      parameters:
      - description: id of transaction
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: transaction data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateTransaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: UPDATE TRANSACTION BY ID
      tags:
      - TRANSACTION
swagger: "2.0"